Significant code cleanup and new features.
The message_replier module is for functions that should trigger when
users talking in chats the bot is in say specific key phrases. Currently
the bot is set up to respond to its name with a random phrase from a
file, complete with string formatting.

Added a /new_sounds command that gives a list of all sounds
added since the last time the command was added.

Added a /delalias command that can be used to delete an existing alias.

Generally gave the bot more personality. Added a 1/1000 chance for the
bot to refuse to play a sound, because that sounded funny at the time.