Implemented chat AI memory & exception logging
The GPT-powered /chat command now has persistent memory to facilitate flowing conversations. The memory has a default limit of 8 messages and can be cleared through the bot with the /lobotomize command.

Exceptions are now logged to a file. This file can be retrieved through the bot with the /logs command.

Also reworked project directory structure to utilize a new Data folder.