Bots now catch token issues and can be individually toggled
The telegram and discord bots can now be toggled on or off with the /setconfig command. This will cause the bot to run or not run the next time the script is started.

Issues with tokens (missing token, invalid token) are now automatically caught and printed to the console without interrupting other bots. Logging can now be toggled on or off.

Added /version and /crash commands (the latter for testing logging support). Added version number to the logs.