Added a command to roll from the d10000 table
Added a /d10000 command that rolls from the Net Libram d10000 table. The bot will track which effects you've rolled, which can be viewed with the /effects command, and reset with the /reseteffects command.

Brought back the username_map.json file to facilitate syncing names between Discord and Telegram for tracking purposes.

Exceptions caught when the bot fails to send a message are now logged.