Added similarity check to sound and search commands
When searching for a sound, sound names with a similarity > 0.75 to the search string will also be accepted, instead of only substrings. This means that poiton will autocorrect to potion, for example.

When requesting a sound to play, if the requested sound name doesn't exist then a search is performed using the same function as the search command. If one result is found, that sound is played. If 2-5 results are found, the results are listed as recommendations. If more than 5 results are found the function returns None like currently.