Bot now always replies to its name
Prewritten responses have been moved to the /test command.

Added an option for a prepended message, which will always be the first non-system message in the bot's memory. This allows you to effectively set the tone for the conversation and is most useful with finetuned models which aren't as responsive to system prompts.