Now checks and assigns superadmin on initial setup
Admins/superadmins, whitelists, and user ID tracking are now platform-specific (discord or telegram).

If a superadmin is not assigned for the platform, the bot will now automatically assign the first person to interact with it. This is only checked ONCE for each platform, though it can be re-enabled using /setconfig. This assists initial setup, as many commands cannot be run without admin or superadmin, and setting these is challenging without already having admin access.

Fixed the discord bot having its user ID tracked.

Fixed the discord bot cutting off the first word of messages when reading messages.

/test command responses can now await functions (necessary now that so many more functions are async)