Replaced exec in /test command with regex string parser
Exec is no longer being used in the /test command for parsing fstrings. Instead, I've written some regex to parse the responses for lists of strings, and randomly choose an element. Also added text replacement tokens for the sender's name and the bot's name for use in responses.

The net effect of this is that ACE is no longer possible through the /test command. This was not a huge issue as you couldn't take advantage of it without file system access anyway, but this does mean that I can now add commands that manipulate the response list without any risk.

Also added docstrings for many UserCommand methods.